Move assertValidationError to HelperAssertion
